requirements.txt
streamlit
google-generativeai
python-dotenv
arxiv
pandas
chromadb
langdetect